Refactor code that generates JSON error messages for standard HTTP error responses
